+++ app/src/androidTest/java/com/github/guliash/calculator/MainActivityTest.java
@@ -5,8 +5,6 @@ import android.support.test.espresso.intent.Intents;
 import android.support.test.espresso.intent.matcher.IntentMatchers;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
-import android.view.View;
-import android.widget.TextView;
 
 import com.guliash.calculator.R;
 import com.guliash.calculator.ui.activities.HelpActivity;
@@ -15,9 +13,6 @@ import com.guliash.calculator.ui.activities.OpenActivity;
 import com.guliash.calculator.ui.activities.SaveActivity;
 import com.guliash.calculator.ui.activities.SettingsActivity;
 
-import org.hamcrest.Description;
-import org.hamcrest.Matcher;
-import org.hamcrest.TypeSafeMatcher;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -30,11 +25,10 @@ import static android.support.test.espresso.action.ViewActions.typeText;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
+import static com.github.guliash.calculator.Matchers.withDoubleText;
 
 @RunWith(AndroidJUnit4.class)
 public class MainActivityTest {
-
-    private static final double EPS = 1e-15;
 
     @Rule
     public ActivityTestRule<MainActivity> mActivityRule = new ActivityTestRule<MainActivity>(
@@ -54,7 +48,7 @@ public class MainActivityTest {
         String stringToTypeIn = "2 + 3";
         onView(withId(R.id.input_field)).perform(typeText(stringToTypeIn));
         onView(withId(R.id.equals_image_button)).perform(click());
-        onView(withId(R.id.result_field)).check(matches(withDouble(5.0)));
+        onView(withId(R.id.result_field)).check(matches(withDoubleText(5.0)));
     }
 
     @Test
@@ -121,26 +115,6 @@ public class MainActivityTest {
         Intents.release();
     }
 
-
-    public static Matcher<View> withDouble(final double expected) {
-        return new TypeSafeMatcher<View>() {
-            @Override
-            protected boolean matchesSafely(View item) {
-                if(!(item instanceof TextView)) {
-                    return false;
-                }
-
-                double actual = Double.valueOf(((TextView) item).getText().toString());
-
-                return Math.abs(actual - expected) < EPS;
-            }
-
-            @Override
-            public void describeTo(Description description) {
-                description.appendText("Actual value differs from expected");
-            }
-        };
-    }
